:recycle: Create a time module.
